Eliminate server crash with FP64 input
Respond with error to non-FP32 input and update README to note that FP32 is the only supported I/O type for the FIL backend at this time.

While attempting to fix the underlying issue here, a double free was revealed that occurred if model execution errored out at a particular point in batch processing. Because the existing implementation was somewhat complicated, it was more difficult to reason about the cause of this double free and ultimately eliminate it. In order to facilitate debugging, the `TRITONBACKEND_ModelInstanceExecute` method was restructured to be more readable and less error-prone.