Eliminate server crash with FP64 input (#119)
* Throw exception if input dtype unexpected

* Eliminate server crash with FP64 input

Respond with error to non-FP32 input and update README to note that FP32 is the only supported I/O type for the FIL backend at this time.

While attempting to fix the underlying issue here, a double free was revealed that occurred if model execution errored out at a particular point in batch processing. Because the existing implementation was somewhat complicated, it was more difficult to reason about the cause of this double free and ultimately eliminate it. In order to facilitate debugging, the `TRITONBACKEND_ModelInstanceExecute` method was restructured to be more readable and less error-prone.

* Update style

* Replace unnecessary decltypes with auto

* Use std::size_t in model_state

* Use AAA style throughout model execution

Ensure that TRITONBACKEND_ModelInstanceExecute adheres to Almost Always
Auto style

* Fix missed AAA opportunities

* Use `make_pair` to generate timing pairs

* Update style

* Additional auto fixes

* Update style