Pin Hypothesis to avoid bug in FloatStrategy
Pin Hypothesis to <6.46.8 to avoid a bug introduced in this release
which prevents float draws being used in test_small. Without this pin,
Hypothesis fails to make the required draw, triggering a bad assert on
`assert smallest_nonzero_magnitude > 0.0` which in turn leads to an
unhandled AttributeError when trying to create a string representation
of the FloatStrategy object since it does not have a min_value
attribute.